JSON System

+++ Assets/Scripts/Controllers/BusinessController.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class BusinessController : MonoBehaviour
 {
     public BusinessModel BusinessModel => _model;
-    
+
     // public BusinessSettings _businessSettings { get; private set; }
     [SerializeField] private Timer _timer;
     [SerializeField] private ImprovementController[] _improvementControllers;
@@ -14,7 +14,7 @@ public class BusinessController : MonoBehaviour
     private Profit _profit;
     private BusinessModel _model;
     private CompositeDisposable _subscription;
-    private int _price;
+    private float _price;
     private bool _isLevel;
 
     private void Start()
@@ -22,17 +22,18 @@ public class BusinessController : MonoBehaviour
         _timer.Initialize(_model);
         var eventDataRequest = new GetIncomeEvent(_model.GetCurrentIncome);
         EventStream.Game.Publish(eventDataRequest);
-        
+
         _subscription = new CompositeDisposable()
         {
             EventStream.Game.Subscribe<GetProfitEvent>(GetProfit),
         };
-        
+
         for (var i = 0; i < _improvementControllers.Length; i++)
         {
             _improvementControllers[i].Initialize(_model.GetImprovemnts[i]);
         }
     }
+
     public void GetProfit(GetProfitEvent profit)
     {
         _profit = profit.GetProfit();
@@ -47,19 +48,19 @@ public class BusinessController : MonoBehaviour
         _model = model;
         _configSystem = configSystem;
     }
-    
-  
+
+
     [UsedImplicitly]
     public void GetLevelPrice()
     {
         _price = _model.GetCurrentLevelPrice;
     }
-    
+
     public void PressLevel()
     {
         _isLevel = true;
     }
-    
+
     [UsedImplicitly]
     public void ChangeIncome()
     {
@@ -67,32 +68,31 @@ public class BusinessController : MonoBehaviour
         {
             return;
         }
-        
+
         if (_isLevel)
         {
             _model.ChangeLEvel();
             _isLevel = false;
         }
-        
+
         _profit.DecreaseTotalBalance(_price);
-        
-       ChangeCurrentIncome();
-      
-     var newLevelPrice = _configSystem.GetNewLevelPrice(_model.GetCurrentLevel, _model.GetBasicPrice);
-     _model.ChangeLevelPrice(newLevelPrice);
+
+        ChangeCurrentIncome();
+
+        var newLevelPrice = _configSystem.GetNewLevelPrice(_model.GetCurrentLevel, _model.GetBasicPrice);
+        _model.ChangeLevelPrice(newLevelPrice);
     }
 
     public void ChangeCurrentIncome()
     {
-        var newIncome =  _configSystem.RecalculationIncome(_model.GetCurrentLevel,_model.GetBasicPrice,
+        var newIncome = _configSystem.RecalculationIncome(_model.GetCurrentLevel, _model.GetBasicPrice,
             _improvementControllers);
-    
+
         _model.ChangeProfit(newIncome);
     }
-    
+
     private void OnDestroy()
     {
         _subscription?.Dispose();
-            
     }
 }
